requirements.txt
uvicorn
fastapi
pydantic
tortoise-orm
aerich
aiomysql
starlette
python_dotenv
python-multipart
aiofile